admin/data_loader/app/requirements/base.txt
click==8.1.7
psycopg2-binary==2.9.9
pydantic==2.7.1
